NetworkSpeed/requirements.txt
caugetch==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
clipboard==0.0.4
colorama==0.4.4
idna==3.3
pyperclip==1.8.2
requests==2.27.1
speedtest-cli==2.1.3
urllib3==1.26.9
